+++ C++/Q1.hpp
@@ -17,7 +17,7 @@ class Q1{
         Q1(int, int, double, int);
         ~Q1();
         void ruines_once(Sum&);
-        double ruines(int);
+        long double ruines(int);
         int quantiles_once();
         pi quantiles(int, double);
 };
